refactor(NavLink): remove the interface to index file
