+++ src/components/navigation/NavLink/index.tsx
@@ -1,7 +1,17 @@
-import { INavLinkProps } from "./interfaces/NavLink.interface";
 import { MdKeyboardArrowRight } from "react-icons/md";
 
 import { StyledNavLink, StyledLink, StyledNavList } from "./styles";
+
+export interface INavLinkProps {
+  id: string;
+  label: string;
+  path: string;
+  isDisabled?: boolean;
+  isSelected?: boolean;
+  icon?: React.ReactNode;
+  handleClick?: () => void;
+  handleBlur?: () => void;
+}
 
 const NavLink = (props: INavLinkProps) => {
   const {
