--- src/components/navigation/NavLink/interfaces/NavLink.interface.ts
@@ -1,10 +1,0 @@
-export interface INavLinkProps {
-  id: string;
-  label: string;
-  path: string;
-  isDisabled?: boolean;
-  isSelected?: boolean;
-  icon?: React.ReactNode;
-  handleClick?: () => void;
-  handleBlur?: () => void;
-}
