+++ src/components/navigation/NavLink/stories/NavLink.Default.stories.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter } from "react-router-dom";
 import { MdHouse } from "react-icons/md";
 
-import { INavLinkProps } from "../interfaces/NavLink.interface";
+import { INavLinkProps } from "..";
 import { NavLinkController } from "./NavLink.Controller";
 import { NavLink } from "..";
